requirements.txt
streamlit
torch
torchvision
transformers
joblib
pandas
numpy
scikit-learn
Pillow
